requirements.txt
amqp==5.0.6
apache-libcloud==3.3.1
asgiref==3.4.1
auto-labeling-pipeline==0.1.21
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.17.104
botocore==1.20.104
bs4==0.0.1
cachetools==4.2.2
celery==5.1.2
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colour==0.1.5
conllu==4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
cycler==0.10.0
daphne==3.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.7
dj-static==0.0.6
Django==3.1.4
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-drf-filepond==0.3.1
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-csv==2.1.1
djangorestframework-xml==2.0.0
djitellopy==1.5
dnspython==1.16.0
doccano==1.4.1
docopt==0.6.2
drf-yasg==1.20.0
ecdsa==0.17.0
environs==9.3.2
et-xmlfile==1.1.0
eventlet==0.31.0
Flask==1.1.2
flatbuffers==1.12
furl==2.1.2
gast==0.3.3
gitignore==0.0.8
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
gunicorn==20.1.0
heroku==0.1.4
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.1
iso8601==0.1.14
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.1.0
lml==0.1.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.12.1
math-function==0.0.1
matplotlib==3.4.1
mplfinance==0.12.7a10
neat==0.4.1
neat-python==0.92
netifaces==0.10.6
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
openpyxl==3.0.7
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==20.9
pandas==1.2.2
pandas-datareader==0.9.0
Pillow==8.1.0
prompt-toolkit==3.0.19
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-xlsx==0.6.0
pygame==2.0.1
PyHamcrest==2.0.2
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypika-tortoise==0.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
s3transfer==0.4.2
scikit-learn==0.24.1
scipy==1.6.0
seqeval==1.2.2
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
socketio==0.2.1
soupsieve==2.2.1
SQLAlchemy==1.4.20
sqlparse==0.4.1
static3==0.7.0
stripe==2.58.0
texttable==1.6.3
threadpoolctl==2.1.0
txaio==20.12.1
typing-extensions==3.7.4.3
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
waitress==2.0.0
wcwidth==0.2.5
WebOb==1.8.7
whitenoise==5.2.0
